refactor: rename Parameters component to ParameterList
